moved the object out of the php file into a seperate file and added a unit test for the object
